2017.11.19.md, 19,695 words. Most prolific day yet.
Wrote on the plane for way less time than I wanted to, then stayed up for 2 20 minute sprints in bed. So about 75 minutes all told and got 3141 words. BALLER.

Figured out so many things about the Delvers and where the Bad Guys come from. This is starting to feel like an actual issue / villain and not just an endless soup of foreboding.  Still have plenty to work out but I'm liking this strategy of just writing the sentence/concept I'm struggling with over and over again with different results. Some of them feel much beter and it's easy to find the righter ones.