Two 30 minute sprints so far today.  I want to do at least one more to get up to ~2700 words for the day so I can start to chip away at my word debt.
Trying to differentiate Lo as a character and not just an avatar of goodness so I trapped us in an elevator. That basically meant I felt really weird about whether or not anybody has jobs in this universe and also I feel like Lo is more likely to ask me a lot of questions than me asking her.  So, in lieu of writing myself into the novel as a character, I stopped the exercise.

Wrote a scene for the baddy.  I think I'm super bored by the prospect of writing all the investigation of the song and that stuff, plus don't know how to keep him front and center for act two.  I think I need some bigger events happening in Act 2.